fix: remove gameLoadFailed call on Script error (this breaks the game launch process)
Script error details - https://ravikiranj.net/posts/2014/code/how-fix-cryptic-script-error-javascript/

When a syntax(?) error occurs in a script, loaded from a different origin, the details of the syntax error are not reported to prevent leaking information (see bug 363897). Instead the error reported is simply "Script error." This behavior can be overriden in some browsers using the crossorigin attribute on and having the server send the appropriate CORS HTTP response headers. A workaround is to isolate "Script error." and handle it knowing that the error detail is only viewable in the browser console and not accessible via JavaScript.